Update invaders based on dojo feedback and testing
