+++ python/pygame_invaders/step_3/config.py
@@ -1,4 +1,4 @@
 ASSET_PATH = "assets/space-shooter-redux/PNG/"
-WIDTH=800
-HEIGHT=600
-FPS=30
+WIDTH = 800
+HEIGHT = 600
+FPS = 30
