+++ python/pygame_invaders/step_4/alien.py
@@ -4,6 +4,7 @@ import pygame
 
 from config import *
 from weapons import AlienBullet
+from explosion import Explosion
 
 
 alien_img = pygame.image.load(ASSET_PATH + "Enemies/enemyRed1.png")
@@ -54,34 +55,10 @@ explosion_anim = [
 ]
 
 
-class AlienExplosion(pygame.sprite.Sprite):
-    def __init__(self, center):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = explosion_anim[0]
-        self.rect = self.image.get_rect()
-        self.rect.center = center
-        self.frame = 0
-        self.last_update = pygame.time.get_ticks()
-        self.frame_rate = 50
-
-    def update(self):
-        now = pygame.time.get_ticks()
-        if now - self.last_update > self.frame_rate:
-            self.last_update = now
-            self.frame += 1
-            if self.frame == len(explosion_anim):
-                self.kill()
-            else:
-                center = self.rect.center
-                self.image = explosion_anim[self.frame]
-                self.rect = self.image.get_rect()
-                self.rect.center = center
-
-
 def collide_aliens(bullets, all_sprites):
     for alien in aliens:
         hits = pygame.sprite.spritecollide(alien, bullets, True)
         if hits:
             alien.kill()
-            explosion = AlienExplosion(alien.rect.center)
+            explosion = Explosion(alien.rect.center, explosion_anim)
             all_sprites.add(explosion)
