+++ python/pygame_invaders/step_5/player.py
@@ -1,5 +1,6 @@
 import pygame
 from config import *
+from explosion import Explosion
 
 player_img = pygame.image.load(ASSET_PATH + "playerShip1_blue.png")
 
@@ -19,33 +20,11 @@ player_explosion_anim = [
 ]
 
 
-class PlayerExplosion(pygame.sprite.Sprite):
-    def __init__(self, center):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = player_explosion_anim[0]
-        self.rect = self.image.get_rect()
-        self.rect.center = center
-        self.frame = 0
-        self.last_update = pygame.time.get_ticks()
-        self.frame_rate = 50
-
-    def update(self):
-        now = pygame.time.get_ticks()
-        if now - self.last_update > self.frame_rate:
-            self.last_update = now
-            self.frame += 1
-            if self.frame == len(player_explosion_anim):
-                self.kill()
-            else:
-                center = self.rect.center
-                self.image = player_explosion_anim[self.frame]
-                self.rect = self.image.get_rect()
-                self.rect.center = center
-
-
 def collide_player_alien_bullets(all_sprites, player, alien_bullets):
+    if not player.alive():
+        return
     hits = pygame.sprite.spritecollide(player, alien_bullets, True)
     if hits:
         player.kill()
-        explosion = PlayerExplosion(player.rect.center)
+        explosion = Explosion(player.rect.center, player_explosion_anim)
         all_sprites.add(explosion)
